add drawer open to more places, project slug
