+++ static/app/views/issueDetails/groupTags/groupTagsDrawer.tsx
@@ -1,15 +1,13 @@
-import {Fragment, useCallback, useRef} from 'react';
+import {useCallback, useRef} from 'react';
 import styled from '@emotion/styled';
 
 import Alert from 'sentry/components/alert';
 import ProjectAvatar from 'sentry/components/avatar/projectAvatar';
 import {Button, LinkButton} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
-import {CompactSelect} from 'sentry/components/compactSelect';
 import Count from 'sentry/components/count';
 import DataExport, {ExportQueryType} from 'sentry/components/dataExport';
 import {DeviceName} from 'sentry/components/deviceName';
-import {DropdownMenu} from 'sentry/components/dropdownMenu';
 import {
   CrumbContainer,
   EventDrawerBody,
@@ -22,245 +20,28 @@ import {
 } from 'sentry/components/events/eventReplay/eventDrawer';
 import {TAGS_DOCS_LINK} from 'sentry/components/events/eventTags/util';
 import useDrawer from 'sentry/components/globalDrawer';
-import GlobalSelectionLink from 'sentry/components/globalSelectionLink';
-import UserBadge from 'sentry/components/idBadge/userBadge';
 import ExternalLink from 'sentry/components/links/externalLink';
 import Link from 'sentry/components/links/link';
 import LoadingError from 'sentry/components/loadingError';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
-import {extractSelectionParameters} from 'sentry/components/organizations/pageFilters/utils';
-import Pagination from 'sentry/components/pagination';
 import Panel from 'sentry/components/panels/panel';
 import PanelBody from 'sentry/components/panels/panelBody';
-import {PanelTable} from 'sentry/components/panels/panelTable';
-import TimeSince from 'sentry/components/timeSince';
 import Version from 'sentry/components/version';
-import {IconArrow, IconEllipsis, IconMail, IconOpen} from 'sentry/icons';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import type {SavedQueryVersions} from 'sentry/types/organization';
 import type {Project} from 'sentry/types/project';
 import {percent} from 'sentry/utils';
-import EventView from 'sentry/utils/discover/eventView';
-import {SavedQueryDatasets} from 'sentry/utils/discover/types';
-import {isUrl} from 'sentry/utils/string/isUrl';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
-import {hasDatasetSelector} from 'sentry/views/dashboards/utils';
+import useProjects from 'sentry/utils/useProjects';
+import {GroupTagsDrawerTagDetails} from 'sentry/views/issueDetails/groupTags/groupTagsDrawerTagDetails';
 import {useGroupTags} from 'sentry/views/issueDetails/groupTags/useGroupTags';
-import {useTagQueries} from 'sentry/views/issueDetails/groupTagValues';
-import {useEnvironmentsFromUrl} from 'sentry/views/issueDetails/utils';
-import {StyledExternalLink} from 'sentry/views/settings/organizationMembers/inviteBanner';
 
 type GroupTagsDrawerProps = {
   groupId: string;
   project: Project;
 };
-
-interface GroupTagsDrawerTagDetailsProps extends GroupTagsDrawerProps {
-  drawerRef: React.RefObject<HTMLDivElement>;
-}
-
-type TagSort = 'date' | 'count';
-const DEFAULT_SORT: TagSort = 'count';
-
-function GroupTagsDrawerTagDetails({groupId, project, drawerRef}: GroupTagsDrawerProps) {
-  const location = useLocation();
-  const organization = useOrganization();
-  const tagKey = location.query.tagDrawerKey as string;
-  const environments = useEnvironmentsFromUrl();
-  const {cursor, page: _page, ...currentQuery} = location.query;
-
-  const title = tagKey === 'user' ? t('Affected Users') : tagKey;
-  const sort: TagSort =
-    (location.query.tagDrawerSort as TagSort | undefined) ?? DEFAULT_SORT;
-  const sortArrow = <IconArrow color="gray300" size="xs" direction="down" />;
-
-  const {tagValueList, tag, isLoading, isError, pageLinks} = useTagQueries({
-    groupId: groupId,
-    sort,
-    tagKey,
-    environments,
-    cursor: typeof cursor === 'string' ? cursor : undefined,
-  });
-
-  const lastSeenColumnHeader = (
-    <StyledSortLink
-      to={{
-        pathname: location.pathname,
-        query: {
-          ...currentQuery,
-          tagDrawerSort: 'date',
-        },
-      }}
-    >
-      {t('Last Seen')} {sort === 'date' && sortArrow}
-    </StyledSortLink>
-  );
-  const countColumnHeader = (
-    <StyledSortLink
-      to={{
-        pathname: location.pathname,
-        query: {
-          ...currentQuery,
-          tagDrawerSort: 'count',
-        },
-      }}
-    >
-      {t('Count')} {sort === 'count' && sortArrow}
-    </StyledSortLink>
-  );
-  const renderResults = () => {
-    if (isError) {
-      return <StyledLoadingError message={t('There was an error loading tag details')} />;
-    }
-
-    if (isLoading) {
-      return null;
-    }
-
-    const discoverFields = [
-      'title',
-      'release',
-      'environment',
-      'user.display',
-      'timestamp',
-    ];
-
-    const globalSelectionParams = extractSelectionParameters(location.query);
-    return tagValueList?.map((tagValue, tagValueIdx) => {
-      const pct = tag?.totalValues
-        ? `${percent(tagValue.count, tag?.totalValues).toFixed(2)}%`
-        : '--';
-      const key = tagValue.key ?? tagKey;
-      const issuesQuery = tagValue.query || `${key}:"${tagValue.value}"`;
-      const discoverView = EventView.fromSavedQuery({
-        id: undefined,
-        name: key ?? '',
-        fields: [
-          ...(key !== undefined ? [key] : []),
-          ...discoverFields.filter(field => field !== key),
-        ],
-        orderby: '-timestamp',
-        // query: `issue:${group.shortId} ${issuesQuery}`,
-        projects: [Number(project?.id)],
-        environment: environments,
-        version: 2 as SavedQueryVersions,
-        range: '90d',
-      });
-      const issuesPath = `/organizations/${organization.slug}/issues/`;
-
-      return (
-        <Fragment key={tagValueIdx}>
-          <NameColumn>
-            <NameWrapper data-test-id="group-tag-value">
-              <GlobalSelectionLink
-                to={{
-                  pathname: `${location.pathname}events/`,
-                  query: {query: issuesQuery},
-                }}
-              >
-                {key === 'user' ? (
-                  <UserBadge
-                    user={{...tagValue, id: tagValue.id ?? ''}}
-                    avatarSize={20}
-                    hideEmail
-                  />
-                ) : (
-                  <DeviceName value={tagValue.name} />
-                )}
-              </GlobalSelectionLink>
-            </NameWrapper>
-
-            {tagValue.email && (
-              <StyledExternalLink
-                href={`mailto:${tagValue.email}`}
-                data-test-id="group-tag-mail"
-              >
-                <IconMail size="xs" color="gray300" />
-              </StyledExternalLink>
-            )}
-            {isUrl(tagValue.value) && (
-              <StyledExternalLink href={tagValue.value} data-test-id="group-tag-url">
-                <IconOpen size="xs" color="gray300" />
-              </StyledExternalLink>
-            )}
-          </NameColumn>
-          <RightAlignColumn>{pct}</RightAlignColumn>
-          <RightAlignColumn>{tagValue.count.toLocaleString()}</RightAlignColumn>
-          <RightAlignColumn>
-            <TimeSince date={tagValue.lastSeen} />
-          </RightAlignColumn>
-          <RightAlignColumn>
-            <DropdownMenu
-              size="sm"
-              position="bottom-end"
-              triggerProps={{
-                size: 'xs',
-                showChevron: false,
-                icon: <IconEllipsis />,
-                'aria-label': t('More'),
-              }}
-              usePortal
-              portalContainerRef={drawerRef}
-              items={[
-                {
-                  key: 'open-in-discover',
-                  label: t('Open in Discover'),
-                  to: discoverView.getResultsViewUrlTarget(
-                    organization.slug,
-                    false,
-                    hasDatasetSelector(organization)
-                      ? SavedQueryDatasets.ERRORS
-                      : undefined
-                  ),
-                  hidden: !organization.features.includes('discover-basic'),
-                },
-                {
-                  key: 'search-issues',
-                  label: t('Search All Issues with Tag Value'),
-                  to: {
-                    pathname: issuesPath,
-                    query: {
-                      ...globalSelectionParams, // preserve page filter selections
-                      query: issuesQuery,
-                    },
-                  },
-                },
-              ]}
-            />
-          </RightAlignColumn>
-        </Fragment>
-      );
-    });
-  };
-
-  return (
-    <Fragment>
-      <StyledPanelTable
-        isLoading={isLoading}
-        isEmpty={!isError && tagValueList?.length === 0}
-        headers={[
-          title,
-          <PercentColumnHeader key="percent">{t('Percent')}</PercentColumnHeader>,
-          countColumnHeader,
-          lastSeenColumnHeader,
-          '',
-        ]}
-        emptyMessage={t('Sorry, the tags for this issue could not be found.')}
-        emptyAction={
-          environments?.length
-            ? t('No tags were found for the currently selected environments')
-            : null
-        }
-      >
-        {renderResults()}
-      </StyledPanelTable>
-      <StyledPagination pageLinks={pageLinks} />
-    </Fragment>
-  );
-}
 
 export function GroupTagsDrawer({project, groupId}: GroupTagsDrawerProps) {
   const location = useLocation();
@@ -326,27 +107,25 @@ export function GroupTagsDrawer({project, groupId}: GroupTagsDrawerProps) {
       <EventNavigator>
         <Header>{tagDrawerKey ? t('Tag Details') : t('Tags')}</Header>
         {tagDrawerKey && (
-          <Fragment>
-            <ButtonBar gap={1}>
-              <LinkButton
-                size="sm"
-                priority="default"
-                href={`/${organization.slug}/${project.slug}/issues/${groupId}/tags/${tagDrawerKey}/export/`}
-              >
-                {t('Export Page to CSV')}
-              </LinkButton>
-              <DataExport
-                payload={{
-                  queryType: ExportQueryType.ISSUES_BY_TAG,
-                  queryInfo: {
-                    project: project.id,
-                    group: groupId,
-                    key: tagDrawerKey,
-                  },
-                }}
-              />
-            </ButtonBar>
-          </Fragment>
+          <ButtonBar gap={1}>
+            <LinkButton
+              size="sm"
+              priority="default"
+              href={`/${organization.slug}/${project.slug}/issues/${groupId}/tags/${tagDrawerKey}/export/`}
+            >
+              {t('Export Page to CSV')}
+            </LinkButton>
+            <DataExport
+              payload={{
+                queryType: ExportQueryType.ISSUES_BY_TAG,
+                queryInfo: {
+                  project: project.id,
+                  group: groupId,
+                  key: tagDrawerKey,
+                },
+              }}
+            />
+          </ButtonBar>
         )}
       </EventNavigator>
       <EventDrawerBody>
@@ -436,23 +215,30 @@ export function GroupTagsDrawer({project, groupId}: GroupTagsDrawerProps) {
 }
 
 export function useGroupTagsDrawer({
-  project,
+  projectSlug,
   groupId,
   openButtonRef,
 }: {
   groupId: string;
   openButtonRef: React.RefObject<HTMLButtonElement>;
-  project: Project;
+  projectSlug: Project['slug'];
 }) {
   const location = useLocation();
   const navigate = useNavigate();
   const drawer = useDrawer();
+  const {projects} = useProjects({slugs: [projectSlug]});
+  const project = projects.find(p => p.slug === projectSlug);
 
   const openTagsDrawer = useCallback(() => {
+    if (!project) {
+      return;
+    }
+
     drawer.openDrawer(() => <GroupTagsDrawer project={project} groupId={groupId} />, {
       ariaLabel: 'tags drawer',
       onClose: () => {
-        if (location.query.tagDrawerSort || location.query.tagDrawerKey) {
+        const params = new URL(window.location.href).searchParams;
+        if (params.has('tagDrawerSort') || params.has('tagDrawerKey')) {
           // Remove drawer state from URL
           navigate(
             {
@@ -476,6 +262,10 @@ export function useGroupTagsDrawer({
       },
     });
   }, [location, navigate, drawer, project, groupId, openButtonRef]);
+
+  if (!project) {
+    return {};
+  }
 
   return {openTagsDrawer};
 }
@@ -562,59 +352,4 @@ const TagBarCount = styled('div')`
   padding-left: ${space(2)};
   padding-right: ${space(1)};
   font-variant-numeric: tabular-nums;
-`;
-
-const StyledPanelTable = styled(PanelTable)`
-  white-space: nowrap;
-  font-size: ${p => p.theme.fontSizeMedium};
-
-  overflow: auto;
-
-  & > * {
-    padding: ${space(1)} ${space(2)};
-  }
-`;
-
-const StyledLoadingError = styled(LoadingError)`
-  grid-column: 1 / -1;
-  margin-bottom: ${space(4)};
-  border-radius: 0;
-  border-width: 1px 0;
-`;
-
-const PercentColumnHeader = styled('div')`
-  text-align: right;
-`;
-
-const StyledSortLink = styled(Link)`
-  text-align: right;
-  color: inherit;
-
-  :hover {
-    color: inherit;
-  }
-`;
-
-const Column = styled('div')`
-  display: flex;
-  align-items: center;
-`;
-
-const NameColumn = styled(Column)`
-  ${p => p.theme.overflowEllipsis};
-  display: flex;
-  min-width: 320px;
-`;
-
-const NameWrapper = styled('span')`
-  ${p => p.theme.overflowEllipsis};
-  width: auto;
-`;
-
-const RightAlignColumn = styled(Column)`
-  justify-content: flex-end;
-`;
-
-const StyledPagination = styled(Pagination)`
-  margin: 0;
 `;
